dbeaver/pro#2309 Fix column resolution order in semantic validation
