+++ plugins/org.jkiss.dbeaver.ui.editors.sql/src/org/jkiss/dbeaver/ui/editors/sql/semantics/model/SQLQueryNodeModelVisitor.java
@@ -47,9 +47,6 @@ public interface SQLQueryNodeModelVisitor<T, R> {
     R visitRowsTableValue(SQLQueryRowsTableValueModel tableValue, T arg);
 
     @Nullable
-    R visitRowsSelectionFilter(SQLQueryRowsSelectionFilterModel selectionFilter, T arg);
-
-    @Nullable
     R visitRowsCrossJoin(SQLQueryRowsCrossJoinModel crossJoin, T arg);
 
     @Nullable
